node routes and model changing

+++ node_routes.js
@@ -1,7 +1,7 @@
 module.exports = function(app, db) {
     app.use(function(req, res, next) {
-        res.header("Access-Control-Allow-Origin", "https://sushibar.herokuapp.com");
         res.header("Access-Control-Allow-Origin", "http://localhost:4200");
+        //res.header("Access-Control-Allow-Origin", "https://sushibar.herokuapp.com");
         res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
         next();
     }),
@@ -13,18 +13,25 @@ module.exports = function(app, db) {
         res.send(products);
     });
     app.post('/goods/create', async (req, res) => {
+        let object = convertToObj(req.body);
+        
+        let price = parseInt(object.price);
+        if (object.name == null || object.description == null || isNaN(price)) return res.send(false);
         let product = await db.Models.Sushi.create({
-            name: req.body.name,
-            description: req.body.description,
-            price: parseInt(req.body.price),
+            name: object.name,
+            description: object.description,
+            price: price,
+            url: object.url,
         });
         res.send(product);
     });
     app.post('/goods/update', async (req, res) => {
+        if (req.body.name == null || req.body.description == null || isNaN(parseInt(req.body.price))) return res.send(false);
         let product = await db.Models.Sushi.update({
             name: req.body.name,
             description: req.body.description,
             price: parseInt(req.body.price),
+            url: req.body.url,
         }, {
             where: {
                 id: parseInt(req.body.id),
@@ -33,6 +40,7 @@ module.exports = function(app, db) {
         res.send(true);
     });
     app.post('/goods/delete', async (req, res) => {
+        if (isNaN(parseInt(req.body.id))) return res.send(false);
         await db.Models.Sushi.destroy({
             where: {
                 id: parseInt(req.body.id),
@@ -41,3 +49,9 @@ module.exports = function(app, db) {
         res.send(true);
     });
 };
+
+let convertToObj = function(obj) {
+    for (const key in obj) {
+        return JSON.parse(key);
+    }
+}
